unittests issues for mpi

+++ tests/dlio_benchmark_test.py
@@ -8,7 +8,8 @@ import hydra
 import unittest
 import yaml
 import shutil
-
+from mpi4py import MPI
+comm = MPI.COMM_WORLD
 import pytest
 
 import os
@@ -17,11 +18,17 @@ from src.dlio_benchmark import DLIOBenchmark
 import glob
 class TestDLIOBenchmark(unittest.TestCase):
     def clean(self) -> None:
-        shutil.rmtree("./checkpoints", ignore_errors=True)
-        shutil.rmtree("./data/", ignore_errors=True)
-        shutil.rmtree("./output", ignore_errors=True)
+        comm.Barrier()
+        if (comm.rank==0):
+            shutil.rmtree("./checkpoints", ignore_errors=True)
+            shutil.rmtree("./data/", ignore_errors=True)
+            shutil.rmtree("./output", ignore_errors=True)
+        comm.Barrier()
     def run_benchmark(self, cfg):
-        shutil.rmtree("./output", ignore_errors=True)
+        comm.Barrier()
+        if (comm.rank==0):
+            shutil.rmtree("./output", ignore_errors=True)
+        comm.Barrier()            
         benchmark = DLIOBenchmark(cfg['workload'])
         benchmark.initialize()
         benchmark.run()
@@ -47,13 +54,13 @@ class TestDLIOBenchmark(unittest.TestCase):
         with initialize(version_base=None, config_path="../configs"):
             cfg = compose(config_name='config', overrides=['++workload.workflow.train=True', '++workload.workflow.generate_data=False', 'workload.train.computation_time=0.01', 'workload.evaluation.eval_time=0.005', 'workload.train.epochs=1'])
             benchmark=self.run_benchmark(cfg)            
-
-            os.makedirs(benchmark.output_folder+"/.hydra/", exist_ok=True)
-            yl: str = OmegaConf.to_yaml(cfg)
-            with open(benchmark.output_folder+"./.hydra/config.yaml", "w") as f:
-                OmegaConf.save(cfg, f)
-            with open(benchmark.output_folder+"./.hydra/overrides.yaml", "w") as f:
-                f.write('[]')
+            if (comm.rank==0):
+                os.makedirs(benchmark.output_folder+"/.hydra/", exist_ok=True)
+                yl: str = OmegaConf.to_yaml(cfg)
+                with open(benchmark.output_folder+"./.hydra/config.yaml", "w") as f:
+                    OmegaConf.save(cfg, f)
+                with open(benchmark.output_folder+"./.hydra/overrides.yaml", "w") as f:
+                    f.write('[]')
             self.assertEqual(len(glob.glob(benchmark.output_folder+"./*_load_and_proc_times.json")), benchmark.comm_size)
     def test2_checkpoint1_epoch(self) -> None:
         with initialize(version_base=None, config_path="../configs"):
@@ -64,7 +71,10 @@ class TestDLIOBenchmark(unittest.TestCase):
                                      '++workload.evaluation.eval_time=0.005', \
                                      '++workload.train.epochs=8', '++workload.workflow.checkpoint=True', \
                                      '++workload.checkpoint.epochs_between_checkpoints=2'])
-            shutil.rmtree("./checkpoints", ignore_errors=True)                                    
+            comm.Barrier()
+            if comm.rank==0:
+                shutil.rmtree("./checkpoints", ignore_errors=True)
+            comm.Barrier()
             benchmark=self.run_benchmark(cfg)            
             self.assertEqual(len(glob.glob("./checkpoints/*.bin")), 4)
     def test2_checkpoint2_step(self) -> None:
@@ -76,12 +86,14 @@ class TestDLIOBenchmark(unittest.TestCase):
                                      '++workload.evaluation.eval_time=0.005', \
                                      '++workload.train.epochs=8', '++workload.workflow.checkpoint=True', \
                                      '++workload.checkpoint.steps_between_checkpoints=2'])
-            benchmark = DLIOBenchmark(cfg['workload'])
+            comm.Barrier()
+            if comm.rank==0:
+                shutil.rmtree("./checkpoints", ignore_errors=True)
+            comm.Barrier()
+            benchmark=self.run_benchmark(cfg)                        
             dataset = cfg['workload']['dataset']
             nstep = dataset.num_files_train * dataset.num_samples_per_file // dataset.batch_size//benchmark.comm_size
             ncheckpoints=nstep//2*8
-            shutil.rmtree("./checkpoints", ignore_errors=True)                        
-            benchmark=self.run_benchmark(cfg)                        
             self.assertEqual(len(glob.glob("./checkpoints/*.bin")), ncheckpoints)
     def test3_eval(self) -> None:
         with initialize(version_base=None, config_path="../configs"):
@@ -102,6 +114,10 @@ class TestDLIOBenchmark(unittest.TestCase):
                                      'workload.train.computation_time=0.01', \
                                      'workload.evaluation.eval_time=0.005', \
                                      '++workload.train.epochs=4', '++workload.workflow.evaluation=True'])
+            comm.Barrier()
+            if comm.rank==0:
+                shutil.rmtree("./output", ignore_errors=True)
+            comm.Barrier()
             benchmark = self.run_benchmark(cfg)                                                
             self.assertEqual(len(glob.glob(benchmark.output_folder+"./*_load_and_proc_times.json")), benchmark.comm_size)            
     def test2_torch_loader(self) -> None:
